+++ Engine/Core/LuaScript/CMakeLists.txt
@@ -1,5 +1,5 @@
 # created on 2021/7/26
 # author @zoloypzuo
-add_subdirectory(LuaBind)
-link_libraries(LuaBind)
 add_library(LuaScript LuaScriptManager.cpp LuaScriptManager.h LuaScriptManager.inl)
+
+target_link_libraries(LuaScript PRIVATE LuaBind)
